Refactor: refactor map help 맵 페이지 도움말 창 리팩토링
