+++ alcoholmap/src/frontend/src/components/units/map/maphelp/Maphelp.types.ts
@@ -1,0 +1,6 @@
+export interface IMapHelpUIProps {
+    helpRef: React.RefObject<HTMLDivElement>;
+    isHelpVisible: boolean;
+    onClickHelpToggleOpen: () => void;
+    onClickCloseHelp: (event: CustomEvent<MouseEvent>) => void;
+}
